requirements.txt
tensorflow
opencv-python
numpy
pandas
streamlit
pillow
pyttsx3
matplotlib
scikit-learn
